+++ src/dot_constants.rs
@@ -45,19 +45,13 @@ pub const EXAMPLE_DOT_OPTIONS: &str = "bgcolor=\"transparent\";\n\tfontsize=50;\
         node [shape=ellipse, penwidth=1, \
         fontname=\"Courier\", pin=true ];\n\tsplines=true;";
 
-/// * activate splines for for `to_dot*`
-///
-/// **Note:** You can chain options by joining the strings
+/// * activate splines
 pub const SPLINES: &str = "splines=true;";
 
-/// * use transparent background for `to_dot*`
-///
-/// **Note:** You can chain options by joining the strings
+/// * use transparent background
 pub const TRANSPARENT_BG: &str = "bgcolor=\"transparent\";";
 
-/// * no overlapping nodes for `to_dot*`
-///
-/// **Note:** You can chain options by joining the strings
+/// * no overlapping nodes
 pub const NO_OVERLAP: &str = "overlap=false;";
 
 /// * Din A4 size